Moved components to separate files, added propTypes to searchBar
